third_party/blink/web_tests/fast/forms/fieldset/legend-margin-with-float-expected.txt

PASS legend.getBoundingClientRect().left is 220
PASS successfullyParsed is true

TEST COMPLETE

Legend
Legend
